Moved aerodynamics to its own folder
