+++ openconcept/aerodynamics/__init__.py
@@ -1,0 +1,1 @@
+from .aerodynamics import PolarDrag, StallSpeed, Lift
