+++ openconcept/analysis/performance/solver_phases.py
@@ -2,7 +2,7 @@ from __future__ import division
 from openmdao.api import Group, ExplicitComponent, IndepVarComp, BalanceComp, ImplicitComponent
 import openconcept.api as oc
 from openconcept.analysis.atmospherics.compute_atmos_props import ComputeAtmosphericProperties
-from openconcept.analysis.aerodynamics import Lift, StallSpeed
+from openconcept.aerodynamics import Lift, StallSpeed
 from openconcept.utilities.math import ElementMultiplyDivideComp, AddSubtractComp
 from openconcept.utilities.math.integrals import Integrator
 from openconcept.utilities.linearinterp import LinearInterpolator
