+++ openconcept/examples/B738_Dymos.py
@@ -4,7 +4,7 @@ import numpy as np
 import openmdao.api as om
 import openconcept.api as oc
 # imports for the airplane model itself
-from openconcept.analysis.aerodynamics import PolarDrag
+from openconcept.aerodynamics import PolarDrag
 from openconcept.examples.aircraft_data.B738 import data as acdata
 from openconcept.analysis.performance.mission_profiles import MissionWithReserve
 from openconcept.components.cfm56 import CFM56
